Change video selection to dmenu
Since there is a dmenu patch to add separators we can change from rofi
to dmenu.
The dmenu patch will just show the video title and output the path/url.

Because we used ' -- ' as a separator and the dmenu option can only
split by a char, we had to change the separator to '`'.